% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rectifySpecialChars}
\alias{rectifySpecialChars}
\title{Rectify special characters.}
\usage{
rectifySpecialChars(x)
}
\arguments{
\item{x}{character vector to work on}
}
\description{
Rectify special characters.
}

