% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R, R/context.R, R/kwic.R, R/read.R, R/dispersion.R
\name{partitionUiInput}
\alias{contextServer}
\alias{contextUiInput}
\alias{contextUiOutput}
\alias{dispersionServer}
\alias{dispersionUiInput}
\alias{dispersionUiOutput}
\alias{kwicServer}
\alias{kwicUiInput}
\alias{kwicUiOutput}
\alias{partitionServer}
\alias{partitionUiInput}
\alias{partitionUiOutput}
\alias{readServer}
\alias{readUiInput}
\alias{readUiOutput}
\title{Building blocks for shiny apps and widgets.}
\usage{
partitionUiInput()

partitionUiOutput()

partitionServer(input, output, session)

contextUiInput()

contextUiOutput()

contextServer(input, output, session)

kwicUiInput(drop = NULL)

kwicUiOutput()

kwicServer(input, output, session, ...)

readUiInput()

readUiOutput()

readServer(input, output, session)

dispersionUiInput()

dispersionUiOutput()

dispersionServer(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{drop}{elements to drop}

\item{...}{further parameters}
}
\description{
Building blocks for shiny apps and widgets.
}

