% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R, R/context.R, R/kwic.R, R/polmineR.R
\docType{methods}
\name{partitionGadget}
\alias{context,missing-method}
\alias{kwic,missing-method}
\alias{partitionGadget}
\alias{polmineR}
\title{Shiny apps and widgets for polmineR.}
\usage{
partitionGadget()

\S4method{context}{missing}()

\S4method{kwic}{missing}()

polmineR()
}
\description{
Shiny apps and widgets for polmineR.
}

