% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R, R/partition.R, R/context.R, R/kwic.R, R/count.R, R/read.R, R/dispersion.R, R/features.R, R/settings.R
\name{corpusUiInput}
\alias{contextServer}
\alias{contextUiInput}
\alias{contextUiOutput}
\alias{corpusServer}
\alias{corpusUiInput}
\alias{corpusUiOutput}
\alias{countServer}
\alias{countUiInput}
\alias{countUiOutput}
\alias{dispersionServer}
\alias{dispersionUiInput}
\alias{dispersionUiOutput}
\alias{featuresServer}
\alias{featuresUiInput}
\alias{featuresUiOutput}
\alias{kwicServer}
\alias{kwicUiInput}
\alias{kwicUiOutput}
\alias{partitionServer}
\alias{partitionUiInput}
\alias{partitionUiOutput}
\alias{readServer}
\alias{readUiInput}
\alias{readUiOutput}
\alias{settingsServer}
\alias{settingsUiInput}
\alias{settingsUiOutput}
\title{Building blocks for shiny apps and widgets.}
\usage{
corpusUiInput()

corpusUiOutput()

corpusServer(input, output, session)

partitionUiInput()

partitionUiOutput()

partitionServer(input, output, session)

contextUiInput()

contextUiOutput()

contextServer(input, output, session)

kwicUiInput(drop = NULL)

kwicUiOutput()

kwicServer(input, output, session, ...)

countUiInput()

countUiOutput()

countServer(input, output, session)

readUiInput()

readUiOutput()

readServer(input, output, session)

dispersionUiInput()

dispersionUiOutput()

dispersionServer(input, output, session)

featuresUiInput()

featuresUiOutput()

featuresServer(input, output, session)

settingsUiInput()

settingsUiOutput()

settingsServer(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{drop}{elements to drop}

\item{...}{further parameters}
}
\description{
Building blocks for shiny apps and widgets.
}

